+++ src/classes/@edl_layer/edl_layer.m
@@ -1,6 +1,0 @@
-classdef edl_layer
-    enumeration
-       diffuse_layer, donnan, no_edl 
-    end
-end
-